Initial e-bike control code
The intial logic to perform a basic e-bike control algorithm on an
Arduino. The algorithm uses 3 LED's, 2 hall effect sensors and 1 ESC.
One hall effect sensor and magnet are placed on the crank and peddles,
the other placed on the back wheel and back forks. The algorithm lights
up one LED when the system is ready, another LED when the crank sensor
is active and the third LED for when the ESC is being powered. The
premise is that as long as the crank sensor has been activated within a
predefined time period, the ESC will slowly ramp up in power, when the
crank sensor has not been activated within the predefined period, the
ESC will then power down.