Incorporate monitoring control system
Incorporate the rotation counting portion of the bike
control system.  Unlike the student control system this
code calculates the speed of both the crank and wheel
to display on the display.  No motor control is included,
when the delta on the crank passes are within the
predefined delay, the system starts counting the wheel
rotations and counts down the remaining time period.